Refactoring of code to make app more modular and easier to scale. Applied bootstrap grid system and styling.
