+++ README.md
@@ -1,1 +1,7 @@
 # brainmaze_server
+
+## install
+`pip install -r requirements.txt`
+
+## run
+`python app.py`
