+++ static/css/menu.css
@@ -1,11 +1,10 @@
 .main-menu {
-    width: 200px;
     background-color: #333;
     color: white;
     height: 100%;
-    position: fixed;
     overflow: auto;
 }
+
 .main-menu ul {
     list-style-type: none;
     padding: 0;
