+++ static/css/styles.css
@@ -12,37 +12,13 @@ body, html {
     background-color: #f1f1f1;
     padding: 1px 10px;
     border-bottom: 1px solid #ccc;
-    position: fixed;
     width: 100%;
     top: 0;
     left: 0;
 }
 
-
 .main-content {
-    /*display: flex;*/
-    margin-top: 3%; /* Adjusted to accommodate the topbar height */
-    margin-left: 200px;
-    height: 97%; /* Adjusted height calculation */
-    flex-direction: column; /* This is crucial for the vertical resizer */,
-
-}
-
-
-main-panel-top, .main-panel-bottom {
-    flex-grow: 1; /* Allows panels to grow and take available space */
-    overflow: hidden; /* Ensures content does not overflow the bounds of the panels */
-}
-
-.main-panel-top {
-    display: flex;
-    flex-direction: row; /* Ensure horizontal layout */
-    height: 70%;
-}
-
-.main-panel-left, .main-panel-right {
-    flex-grow: 1; /* Ensure panels grow equally */
-    min-width: 100px; /* Minimum width to prevent panels from disappearing */
+    height: 100%; 
 }
 
 .main-panel-left {
@@ -52,7 +28,6 @@ main-panel-top, .main-panel-bottom {
     background-color: #00f9f9;
     overflow-x: auto; /* Enables horizontal scrolling if needed */
     overflow-y: auto;
-    height: 100%; /* Sets the height to fill the container */
 }
 
 
@@ -72,9 +47,7 @@ main-panel-top, .main-panel-bottom {
 
 
 .main-panel-bottom {
-    display: flex;
     flex-direction: column;
-    align-items: stretch; /* Align children to stretch across the width */
     width: 100%; /* Ensure it takes the full width of its parent */
     height: auto; /* Adjust height as needed */
 }
@@ -85,9 +58,20 @@ main-panel-top, .main-panel-bottom {
 }
 
 #eeg-slider {
-    width: 100%; /* Slider takes the full width */
+    width: 98%; /* Slider takes the full width */
 }
 
 .selected {
     background-color: #d9edf7;  /* Light blue color, adjust as needed */
+}
+
+/* DataTables styling */
+.dataTables_wrapper {
+    height: 700px; /* Set the desired height */
+    overflow-y: auto; /* Add vertical scrollbar if needed */
+}
+
+.dataTables_wrapper table.dataTable tbody th,
+.dataTables_wrapper table.dataTable tbody td {
+    font-size: 16px; /* Set the desired font size */
 }
