+++ templates/login.html
@@ -2,13 +2,16 @@
 <html>
 <head>
     <title>Register</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
 </head>
-<body>
-    <h2>LOGIN</h2>
-    <form action="" method="post">
-        Username: <input type="text" name="username" required><br>
-        Password: <input type="password" name="password" required><br>
-        <input type="submit" value="Loginr">
-    </form>
+<body class="bg-secondary">
+    <div class="container mt-5">
+        <h2>LOGIN</h2>
+        <form action="" method="post" class="p-3">
+            Username: <input type="text" name="username" required class="form-control form-control-sm"><br>
+            Password: <input type="password" name="password" required class="form-control form-control-sm"><br>
+            <input type="submit" value="Login" class="btn btn-primary">
+        </form>
+    </div>
 </body>
 </html>
